pyproject.toml
[project]
name = "simple-fluid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.5",
    "pyglet>=2.1.8",
    "usd-core>=25.8",
    "warp-lang>=1.8.1",
]
